refactor: Entity와 Enum Class 분리
