+++ src/main/java/Enum/Status.java
@@ -1,4 +1,4 @@
-package Entity;
+package Enum;
 
 public enum Status {
     GAME_START("숫자 야구 게임을 시작합니다."),
